+++ cmd/generate/scaffold.go
@@ -96,7 +96,6 @@ var ScaffoldCmd = &cobra.Command{
 			Path:                   "models",
 			Package:                "models",
 			TestPackage:            "models",
-			Encoding:               "json",
 			ForceDefaultID:         true,
 			ForceDefaultTimestamps: true,
 		}